requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
diff-match-patch==20230430
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.1.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.15.4
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.3
whitenoise==6.7.0
